refactor(lp): remove old lp artifacts and remove failed tx dumps
