--- contracts/liquidity-pool-typescript-lp/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "liquidity-pool-typescript-lp"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ckb-std = "0.17.0"
-perun-common = { path = "../../crates/perun-common", default-features = false, features = ["contract"] }
-molecule = { version = "0.8.0", default-features = false, features = ["bytes_vec"] }
-
-[features]
-library = []
-native-simulator = ["library", "ckb-std/native-simulator"]
