--- contracts/liquidity-pool-typescript-lp/src/lib.rs
@@ -1,11 +1,0 @@
-#![cfg_attr(not(feature = "library"), no_std)]
-#![allow(special_module_name)]
-#![allow(unused_attributes)]
-
-#[cfg(feature = "library")]
-extern crate alloc;
-
-#[cfg(feature = "library")]
-mod main;
-#[cfg(feature = "library")]
-pub use main::program_entry;
