--- contracts/liquidity-pool-typescript/.gitignore
@@ -1,2 +1,0 @@
-/build
-/target
