+++ contracts/liquidity-pool-typescript/Makefile
@@ -3,29 +3,14 @@
 cur_dir = $(dir $(abspath $(firstword $(MAKEFILE_LIST))))
 
 TOP := $(cur_dir)
-# RUSTFLAGS that are likely to be tweaked by developers. For example,
-# while we enable debug logs by default here, some might want to strip them
-# for minimal code size / consumed cycles.
 CUSTOM_RUSTFLAGS := -C debug-assertions
-# RUSTFLAGS that are less likely to be tweaked by developers. Most likely
-# one would want to keep the default values here.
 FULL_RUSTFLAGS := -C target-feature=+zba,+zbb,+zbc,+zbs,-a $(CUSTOM_RUSTFLAGS)
-# Additional cargo args to append here. For example, one can use
-# make test CARGO_ARGS="-- --nocapture" so as to inspect data emitted to
-# stdout in unit tests
 CARGO_ARGS :=
 MODE := release
-# Tweak this to change the clang version to use for building C code. By default
-# we use a bash script with some heuristics to find clang in current system.
 CLANG := $(shell $(TOP)/scripts/find_clang)
 AR := $(subst clang,llvm-ar,$(CLANG))
 OBJCOPY := $(subst clang,llvm-objcopy,$(CLANG))
-# When this is set to some value, the generated binaries will be copied over
 BUILD_DIR :=
-# Generated binaries to copy. By convention, a Rust crate's directory name will
-# likely match the crate name, which is also the name of the final binary.
-# However if this is not the case, you can tweak this variable. As the name hints,
-# more than one binary is supported here.
 BINARIES := $(notdir $(shell pwd))
 
 ifeq (release,$(MODE))
@@ -47,8 +32,6 @@ build:
 		done \
 	fi
 
-# test, check, clippy and fmt here are provided for completeness,
-# there is nothing wrong invoking cargo directly instead of make.
 test:
 	cargo test $(CARGO_ARGS)
 
@@ -61,12 +44,6 @@ clippy:
 fmt:
 	cargo fmt $(CARGO_ARGS)
 
-# Arbitrary cargo command is supported here. For example:
-#
-# make cargo CARGO_CMD=expand CARGO_ARGS="--ugly"
-#
-# Invokes:
-# cargo expand --ugly
 CARGO_CMD :=
 cargo:
 	cargo $(CARGO_CMD) $(CARGO_ARGS)
