--- contracts/liquidity-pool-typescript/README.md
@@ -1,7 +1,0 @@
-# liquidity-pool-typescript
-
-TODO: Write this readme
-
-*This contract was bootstrapped with [ckb-script-templates].*
-
-[ckb-script-templates]: https://github.com/cryptape/ckb-script-templates
