--- crates/perun-common/liquidity_pool_lp.mol
@@ -1,60 +1,0 @@
-import blockchain;
-
-array PoolId [byte; 32];
-
-table LPPolicy {
-    max_trading_volume: Uint64,
-    fee_rate_bps:       Uint32,
-    policy_flags:       Uint32,
-    policy_version:     Uint32,
-}
-
-table LPCell {
-    pool_id:                    PoolId,
-    owner_lock_hash:            Byte32,
-    operator_lock_hash:         Byte32,
-    available_ckb:              Uint64,
-    reserved_ckb:               Uint64,
-    cumulative_fees_earned_ckb: Uint64,
-    policy:                     LPPolicy,
-    nonce:                      Uint64,
-    active:                     byte,
-}
-
-table LPDepositWitness {}
-
-table LPWithdrawWitness {
-    ckb_out: Uint64,
-}
-
-table FundChannelExtractWitness {
-    channel_id:      Byte32,
-    contribution_id: Byte32,
-    extract_ckb:     Uint64,
-}
-
-table SettleChannelInsertWitness {
-    channel_id:         Byte32,
-    contribution_id:    Byte32,
-    principal_returned: Uint64,
-    fee_ckb:            Uint64,
-    price_x64:          Uint128,
-}
-
-table CancelReservationWitness {
-    channel_id:      Byte32,
-    contribution_id: Byte32,
-}
-
-table RotateOperatorWitness {
-    new_operator_lock_hash: Byte32,
-}
-
-union PoolWitness {
-    LPDepositWitness,
-    LPWithdrawWitness,
-    FundChannelExtractWitness,
-    SettleChannelInsertWitness,
-    CancelReservationWitness,
-    RotateOperatorWitness,
-}
